# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vera Nexus BI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /thoughts:
    get:
      operationId: listThoughts
      tags: [thoughts]
      summary: List thoughts
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, technology, finance, markets, health]
      responses:
        "200":
          description: List of thoughts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Thought"

    post:
      operationId: createThought
      tags: [thoughts]
      summary: Post a new thought
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThoughtInput"
      responses:
        "201":
          description: Created thought
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thought"
  /thoughts/{id}:
    get:
      operationId: getThought
      tags: [thoughts]
      summary: Get a single thought
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Thought
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Thought"

    delete:
      operationId: deleteThought
      tags: [thoughts]
      summary: Delete a thought
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── REACTIONS ──
  /reactions:
    post:
      operationId: toggleReaction
      tags: [reactions]
      summary: Toggle a reaction on a thought
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Reaction toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResult"

  # ── VENUS AI ──
...
  /settings/groq-key:
    get:
      operationId: getGroqKeyStatus
      tags: [settings]
      summary: Check if Groq API key is configured
      responses:
        "200":
          description: Key status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroqKeyStatus"

    post:
      operationId: saveGroqKey
      tags: [settings]
      summary: Save Groq API key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroqKeyInput"
      responses:
        "200":
          description: Key saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroqKeyStatus"

    delete:
      operationId: deleteGroqKey
      tags: [settings]
      summary: Remove Groq API key
      responses:
        "200":
          description: Key removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroqKeyStatus"
  /settings/onboarding:
    get:
      operationId: getOnboarding
      tags: [settings]
      summary: Get onboarding/business context
      responses:
        "200":
          description: Onboarding data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingData"

    post:
      operationId: saveOnboarding
      tags: [settings]
      summary: Save onboarding/business context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingInput"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingData"
components:
  schemas:
...
    Thought:
      type: object
      required: [id, content, author, category, createdAt]
      properties:
        id:
          type: integer
        content:
          type: string
        author:
          type: string
        category:
          type: string
          enum: [technology, finance, markets, health]
        createdAt:
          type: string
        reactions:
          type: object
          additionalProperties:
            type: integer
        tags:
          type: array
          items:
            type: string
    ThoughtInput:
      type: object
      required: [content, author, category]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 500
        author:
          type: string
          minLength: 1
        category:
          type: string
          enum: [technology, finance, markets, health]
        tags:
          type: array
          items:
            type: string

    # ── REACTION ──
    ReactionInput:
      type: object
      required: [thoughtId, reactionType]
      properties:
        thoughtId:
          type: integer
        reactionType:
          type: string
          enum: [fire, brain, chart, signal, agree]
    ReactionResult:
      type: object
      required: [action, thoughtId, reactionType]
      properties:
        action:
          type: string
          enum: [added, deleted]
        thoughtId:
          type: integer
        reactionType:
          type: string

    # ── VENUS AI ──
...
    GroqKeyStatus:
      type: object
      required: [configured]
      properties:
        configured:
          type: boolean
        maskedKey:
          type: ["string", "null"]
    GroqKeyInput:
      type: object
      required: [apiKey]
      properties:
        apiKey:
          type: string
          minLength: 10
    OnboardingData:
      type: object
      properties:
        companyName:
          type: ["string", "null"]
        stage:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        teamSize:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        primaryGoal:
          type: ["string", "null"]
        completed:
          type: boolean
    OnboardingInput:
      type: object
      required: [companyName, stage, industry, teamSize, country, primaryGoal]
      properties:
        companyName:
          type: string
        stage:
          type: string
        industry:
          type: string
        teamSize:
          type: string
        country:
          type: string
        primaryGoal:
          type: string