# lib/api-zod/src/generated/types/thought.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Vera Nexus BI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ThoughtCategory } from './thoughtCategory';
import type { ThoughtReactions } from './thoughtReactions';

export interface Thought {
  id: number;
  content: string;
  author: string;
  category: ThoughtCategory;
  createdAt: string;
  reactions?: ThoughtReactions;
  tags?: string[];
}